move configurations to constants
